+++ fixtures/app_with_spm_dependencies/Features/FeatureOne/Sources/ObjectiveFeature.h
@@ -1,0 +1,16 @@
+//
+//  ObjectiveFeature.h
+//  FeatureOne
+//
+//  Created by Shai Mishali on 23/03/2024.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface ObjectiveFeature : NSObject
+
+@end
+
+NS_ASSUME_NONNULL_END