+++ fixtures/multiplatform_app_with_sdk/Tuist/Package.swift
@@ -4,6 +4,6 @@ import PackageDescription
 let package = Package(
     name: "PackageName",
     dependencies: [
-        .package(url: "https://github.com/firebase/firebase-ios-sdk", from: "10.15.0"),
+        .package(url: "https://github.com/firebase/firebase-ios-sdk", from: "10.23.0"),
     ]
 )