+++ fixtures/multiplatform_µFeature_unit_tests_with_explicit_dependencies/Modules/ModuleA/Example/Sources/AppDelegate.swift
@@ -1,0 +1,11 @@
+import ModuleAImplementation
+import SwiftUI
+
+@main
+struct TuistApp: App {
+    var body: some Scene {
+        WindowGroup {
+            Text("Tuist is great")
+        }
+    }
+}