+++ fixtures/multiplatform_µFeature_unit_tests_with_explicit_dependencies/Modules/ModuleA/TestSupporting/Sources/FrameworkATestSupporting.swift
@@ -1,0 +1,13 @@
+import Foundation
+import Mocker
+
+public enum NetworkResponseMocks {
+    public static var testMock: Mock {
+        Mock(
+            url: URL(string: "https://apple.com")!,
+            dataType: .json,
+            statusCode: 200,
+            data: [.get: Data()]
+        )
+    }
+}