+++ fixtures/multiplatform_µFeature_unit_tests_with_explicit_dependencies/Tuist/Config.swift
@@ -1,0 +1,3 @@
+import ProjectDescription
+
+let config = Config(generationOptions: .options(enforceExplicitDependencies: true))