+++ fixtures/multiplatform_µFeature_unit_tests_with_explicit_dependencies/Tuist/Dependencies.swift
@@ -1,0 +1,11 @@
+import ProjectDescription
+import ProjectDescriptionHelpers
+
+let dependencies = Dependencies(
+    swiftPackageManager: .init(
+        productTypes: [
+            "Mocker": .framework,
+        ]
+    ),
+    platforms: [.iOS, .watchOS]
+)